+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerHeaderFragment.kt
@@ -53,6 +53,7 @@ import coil.load
 import coil.request.Disposable
 import coil.request.ErrorResult
 import coil.request.ImageRequest
+import coil.size.Scale
 import coil.size.Size
 import coil.transform.RoundedCornersTransformation
 import com.airbnb.lottie.LottieAnimationView
@@ -437,7 +438,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
 
             val imageBuilder: ImageRequest.Builder.() -> Unit = {
                 error(IR.drawable.defaultartwork_dark)
-                size(Size.ORIGINAL)
+                scale(Scale.FIT)
                 transformations(RoundedCornersTransformation(imageLoader.radiusPx.toFloat()), ThemedImageTintTransformation(imageView.context))
             }