+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/components/UserEpisodeImage.kt
@@ -24,6 +24,7 @@ fun UserEpisodeImage(
             .build(),
         contentDescription = contentDescription,
         placeholder = painterResource(placeholder),
+        fallback = painterResource(placeholder),
         modifier = modifier,
     )
 }