+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/player/EffectsScreen.kt
@@ -16,7 +16,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -27,10 +26,11 @@ import androidx.wear.compose.material.Icon
 import androidx.wear.compose.material.InlineSlider
 import androidx.wear.compose.material.InlineSliderDefaults
 import androidx.wear.compose.material.MaterialTheme
+import androidx.wear.tooling.preview.devices.WearDevices
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH30
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH50
-import au.com.shiftyjelly.pocketcasts.models.to.PlaybackEffects
+import au.com.shiftyjelly.pocketcasts.models.to.PlaybackEffectsData
 import au.com.shiftyjelly.pocketcasts.models.type.TrimMode
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.wear.ui.component.ScreenHeaderChip
@@ -236,7 +236,7 @@ fun TrimSilenceSlider(
 }
 
 @Preview(
-    device = Devices.WEAR_OS_LARGE_ROUND,
+    device = WearDevices.LARGE_ROUND,
     showSystemUi = true,
     backgroundColor = 0xff000000,
     showBackground = true
@@ -247,11 +247,11 @@ private fun EffectsScreenDarkPreview() {
         Content(
             columnState = belowTimeTextPreview(),
             state = EffectsViewModel.State.Loaded(
-                playbackEffects = PlaybackEffects().apply {
-                    trimMode = TrimMode.MEDIUM
-                    playbackSpeed = 1.5
-                    isVolumeBoosted = true
-                }
+                playbackEffects = PlaybackEffectsData(
+                    trimMode = TrimMode.MEDIUM,
+                    playbackSpeed = 1.5,
+                    isVolumeBoosted = true,
+                )
             ),
             increasePlaybackSpeed = {},
             decreasePlaybackSpeed = {},